feat: transition to monorepo to be able to run example app with wallet plugin from the same repo
