--- examples/react/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "web3js-react-dapp-demo",
-  "version": "0.1.0",
-  "description": "Demonstration Web3.js + React dApp",
-  "contributors": [
-    "ChainSafe <info@chainsafe.io>"
-  ],
-  "license": "MIT",
-  "dependencies": {
-    "@types/node": "^16.18.108",
-    "@types/react": "^18.3.8",
-    "@types/react-dom": "^18.3.0",
-    "@types/uuid": "^10.0.0",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.9.5",
-    "uuid": "^10.0.0",
-    "web3": "^4.13.0",
-    "web3-eth": "^4.9.0"
-  },
-  "scripts": {
-    "start": "react-scripts start"
-  }
-}
