+++ package.json
@@ -1,17 +1,11 @@
 {
   "name": "web3-plugin-wallet-rpc",
+  "private": true,
   "version": "1.0.0",
   "description": "Web3.js plugin to add support for wallet-related RPC methods",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
   "homepage": "https://github.com/web3/web3-wallet-rpc-utils#readme",
   "bugs": {
     "url": "https://github.com/web3/web3-wallet-rpc-utils/issues"
-  },
-  "scripts": {
-    "lint": "eslint '{src,test}/**/*.ts'",
-    "build": "tsc --project tsconfig.build.json",
-    "test": "jest --config=./test/jest.config.js"
   },
   "contributors": [
     "ChainSafe <info@chainsafe.io>"
@@ -21,20 +15,15 @@
     "type": "git",
     "url": "git@github.com:web3/web3-wallet-rpc-utils.git"
   },
-  "dependencies": {},
-  "devDependencies": {
-    "@chainsafe/eslint-config": "^2.0.0",
-    "@types/jest": "^29.5.2",
-    "@types/node": "^20.2.6",
-    "eslint": "8",
-    "jest": "^29.5.0",
-    "jest-extended": "^4.0.0",
-    "ts-jest": "^29.1.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.3",
-    "web3": "^4.2.2"
-  },
-  "peerDependencies": {
-    "web3": ">= 4.0.3"
+  "workspaces": [
+    "packages/example-react-app",
+    "packages/web3-plugin-wallet-rpc"
+  ],
+  "packageManager": "yarn@1.22.15",
+  "scripts": {
+    "lint": "cd packages/web3-plugin-wallet-rpc && yarn lint",
+    "build": "cd packages/web3-plugin-wallet-rpc && yarn build",
+    "test": "cd packages/web3-plugin-wallet-rpc && yarn test",
+    "start": "cd packages/example-react-app && yarn start"
   }
 }
