+++ packages/example-react-app/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "example-react-app",
+  "private": true,
+  "version": "0.1.0",
+  "description": "Demonstration Web3.js + React dApp",
+  "contributors": [
+    "ChainSafe <info@chainsafe.io>"
+  ],
+  "license": "MIT",
+  "dependencies": {
+    "@types/node": "^16.18.108",
+    "@types/react": "^18.3.8",
+    "@types/react-dom": "^18.3.0",
+    "@types/uuid": "^10.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-scripts": "5.0.1",
+    "typescript": "5.3.x",
+    "uuid": "^10.0.0",
+    "web3": "^4.13.0",
+    "web3-eth": "^4.9.0",
+    "web3-plugin-wallet-rpc": "*"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "lint": "eslint 'src/**/*.{ts,tsx}'"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
