+++ packages/example-react-app/public/favicon.ico
