+++ packages/example-react-app/public/manifest.json
