+++ packages/web3-plugin-wallet-rpc/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: "@chainsafe",
+};
