+++ packages/web3-plugin-wallet-rpc/.gitignore
@@ -1,1 +1,2 @@
 node_modules
+lib
