+++ packages/web3-plugin-wallet-rpc/README.md
@@ -1,0 +1,3 @@
+# Web3.js Plugin for Wallet RPC methods
+
+Refer to the README file located in the root of this repository.
