+++ packages/web3-plugin-wallet-rpc/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "web3-plugin-wallet-rpc",
+  "version": "1.0.0",
+  "description": "Add support for the wallet-related RPC methods",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "homepage": "https://github.com/web3/web3-wallet-rpc-utils#readme",
+  "bugs": {
+    "url": "https://github.com/web3/web3-wallet-rpc-utils/issues"
+  },
+  "scripts": {
+    "lint": "eslint '{src,test}/**/*.ts'",
+    "build": "tsc --project tsconfig.build.json",
+    "test": "jest --config=./test/jest.config.js"
+  },
+  "contributors": [
+    "ChainSafe <info@chainsafe.io>"
+  ],
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:web3/web3-wallet-rpc-utils.git"
+  },
+  "devDependencies": {
+    "@chainsafe/eslint-config": "^2.0.0",
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.2.6",
+    "eslint": "8",
+    "jest": "^29.5.0",
+    "jest-extended": "^4.0.0",
+    "ts-jest": "^29.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "5.3.x",
+    "web3": "^4.2.2"
+  },
+  "peerDependencies": {
+    "web3": ">= 4.0.3"
+  }
+}
