+++ packages/web3-plugin-wallet-rpc/src/utils.ts
@@ -1,4 +1,4 @@
-import { OwnedAsset } from "./types";
+import type { OwnedAsset } from "./types";
 
 export function parseToGetOwnedAssetsResult(result: unknown[]): OwnedAsset[] {
   return result as OwnedAsset[];
