+++ packages/web3-plugin-wallet-rpc/test/WalletRpcPlugin.test.ts
@@ -1,4 +1,5 @@
 import { core } from "web3";
+
 import { WalletRpcPlugin } from "../src";
 
 describe("WalletRpcPlugin", () => {
