+++ packages/web3-plugin-wallet-rpc/test/wallet_addEthereumChain.test.ts
@@ -1,4 +1,5 @@
 import { Web3 } from "web3";
+
 import { WalletRpcPlugin } from "../src";
 
 describe("WalletRpcPlugin", () => {
