+++ packages/web3-plugin-wallet-rpc/test/wallet_switchEthereumChain.test.ts
@@ -1,8 +1,9 @@
 import { Web3 } from "web3";
+
 import { WalletRpcPlugin } from "../src";
 
 describe("WalletRpcPlugin", () => {
-  describe("wallet_getOwnedAssets", () => {
+  describe("wallet_switchEthereumChain", () => {
     const web3 = new Web3("http://127.0.0.1:8545");
     web3.registerPlugin(new WalletRpcPlugin());
 
@@ -14,20 +15,18 @@ describe("WalletRpcPlugin", () => {
     });
 
     it("should call the method with expected params", async () => {
-      await web3.walletRpc.getOwnedAssets({
-        address: "0xa5653e88D9c352387deDdC79bcf99f0ada62e9c6",
-      });
+      await web3.walletRpc.switchEthereumChain(5000);
 
       expect(requestManagerSendSpy).toHaveBeenCalledWith({
-        method: "wallet_getOwnedAssets",
-        params: [{ address: "0xa5653e88D9c352387deDdC79bcf99f0ada62e9c6" }],
+        method: "wallet_switchEthereumChain",
+        params: [{ chainId: "0x1388" }],
       });
     });
 
-    it("should throw when called with invalid address", async () => {
-      await expect(
-        web3.walletRpc.getOwnedAssets({ address: "" })
-      ).rejects.toThrow("validator found 1 error");
+    it("should return correct result", async () => {
+      const result = await web3.walletRpc.switchEthereumChain(5000);
+
+      expect(result).toBeUndefined();
     });
   });
 });
