+++ packages/web3-plugin-wallet-rpc/test/wallet_watchAsset.test.ts
@@ -1,8 +1,9 @@
 import { Web3 } from "web3";
+
 import { WalletRpcPlugin } from "../src";
 
 describe("WalletRpcPlugin", () => {
-  describe("wallet_switchEthereumChain", () => {
+  describe("wallet_watchAsset", () => {
     const web3 = new Web3("http://127.0.0.1:8545");
     web3.registerPlugin(new WalletRpcPlugin());
 
@@ -14,16 +15,29 @@ describe("WalletRpcPlugin", () => {
     });
 
     it("should call the method with expected params", async () => {
-      await web3.walletRpc.switchEthereumChain(5000);
+      const request = {
+        type: "ERC20",
+        options: {
+          address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+          symbol: "USDC",
+        },
+      };
+
+      await web3.walletRpc.watchAsset(request);
 
       expect(requestManagerSendSpy).toHaveBeenCalledWith({
-        method: "wallet_switchEthereumChain",
-        params: [{ chainId: "0x1388" }],
+        method: "wallet_watchAsset",
+        params: [request],
       });
     });
 
     it("should return correct result", async () => {
-      const result = await web3.walletRpc.switchEthereumChain(5000);
+      const result = await web3.walletRpc.watchAsset({
+        type: "ERC20",
+        options: {
+          address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
+        },
+      });
 
       expect(result).toBeUndefined();
     });
